Office Module Authorized !!

+++ app/Http/Controllers/OfficeController.php
@@ -10,31 +10,39 @@ class OfficeController extends Controller
 {
     public function index()
     {
-        $offices = Office::get()->all();
-        // return response()->json($offices);
-        return view('hrm.office.index', compact('offices'));
+        if (auth()->user()->can('office_view_all')) {
+            $offices = Office::get()->all();
+            return view('hrm.office.index', compact('offices'));
+        }
+        return abort('403', __('You are not authorized'));
     }
     public function create()
     {
-        $companies = DB::table('company')->get();
-        return view('hrm.office.create', compact('companies'));
+        if (auth()->user()->can('office_create')) {
+            $companies = DB::table('company')->get();
+            return view('hrm.office.create', compact('companies'));
+        }
+        return abort('403', __('You are not authorized'));
     }
 
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
-            'name' => 'required|string|max:255',
-            'clock_in' => 'required|string',
-            'clock_out' => 'required|string',
-            'company' => 'required|exists:company,id',
-        ]);
-        $office = new Office();
-        $office->name = $validatedData['name'];
-        $office->clock_in = $validatedData['clock_in'];
-        $office->clock_out = $validatedData['clock_out'];
-        $office->company_id = $validatedData['company'];
-        $office->save();
-        return redirect(route('office.index'))->with('success', 'Office created successfully!');
+        if (auth()->user()->can('office_create')) {
+            $validatedData = $request->validate([
+                'name' => 'required|string|max:255',
+                'clock_in' => 'required|string',
+                'clock_out' => 'required|string',
+                'company' => 'required|exists:company,id',
+            ]);
+            $office = new Office();
+            $office->name = $validatedData['name'];
+            $office->clock_in = $validatedData['clock_in'];
+            $office->clock_out = $validatedData['clock_out'];
+            $office->company_id = $validatedData['company'];
+            $office->save();
+            return redirect(route('office.index'))->with('success', 'Office created successfully!');
+        }
+        return abort('403', __('You are not authorized'));
     }
     public function printData()
     {
@@ -43,37 +51,42 @@ class OfficeController extends Controller
     }
     public function delete(Request $request)
     {
-        $office = Office::findOrFail($request->id);
-    
-        if ($office) {
-            $office->delete();
-            return response()->json(['message' => 'Office Shift deleted successfully'], 200);
-        } else {
-            return response()->json(['message' => 'Office Shift not found'], 404);
+        if (auth()->user()->can('office_delete')) {
+            $office = Office::findOrFail($request->id);
+            if ($office) {
+                $office->delete();
+                return response()->json(['message' => 'Office Shift deleted successfully'], 200);
+            } else {
+                return response()->json(['message' => 'Office Shift not found'], 404);
+            }
         }
+        return abort('403', __('You are not authorized'));
     }
 
     public function edit($id)
     {
-        $companies = DB::table('company')->get();
-        $office = Office::findOrFail($id);
-        return view('hrm.office.edit', compact('office', 'companies'));
+        if (auth()->user()->can('office_edit')) {
+            $companies = DB::table('company')->get();
+            $office = Office::findOrFail($id);
+            return view('hrm.office.edit', compact('office', 'companies'));
+        }
+        return abort('403', __('You are not authorized'));
     }
 
     public function update(Request $request, $id)
-{
-    $request->validate([
-        'name' => 'required|string|max:255',
-        'clock_in' => 'required|date_format:H:i',
-        'clock_out' => 'required',
-        'company' => 'required|exists:company,id',
-    ]);
+    {
+        if (auth()->user()->can('office_edit')) {
+            $request->validate([
+                'name' => 'required|string|max:255',
+                'clock_in' => 'required|date_format:H:i',
+                'clock_out' => 'required',
+                'company' => 'required|exists:company,id',
+            ]);
+            $office = Office::findOrFail($id);
+            $office->update($request->all());
 
-    $office = Office::findOrFail($id);
-    $office->update($request->all());
-
-    return redirect()->route('office.index')->with('success', 'Office updated successfully');
-}
-
-
+            return redirect()->route('office.index')->with('success', 'Office updated successfully');
+        }
+        return abort('403', __('You are not authorized'));
+    }
 }
